Improvised Tools (#26)
* adding my crappy sprites 🐕

* putting all of ze code into a file

* adds custom fuel types on welders

in case you want to certain welders to be refueled with different liquids... such as beer >:)

* adds a var that allows welders to not be turned off

this will mean that they will just keep running until there is no more fuel left! of course, this won't effect any welder except the improvised welder.

* ough

* oopz for got to reference icon file

* makes improv crowbar able to break

adds a durability variable to the improvised crowbar, where if it finishes performing certain actions, it'll deduct durability. Buffs the speed to be closer to default to make it not totally horrible

---------

Co-authored-by: NoeSilent <111362881+NoeSilent@users.noreply.github.com>